requirements.txt
matplotlib==3.5.2
pandas==1.4.4
Pillow==9.2.0
Pillow==10.3.0
plotly==5.9.0
streamlit==1.33.0
